all profile is runnable

+++ clSandbox/kernel.cl
@@ -68,7 +68,7 @@ __kernel void computeStats(__global double* buffer, uint64_t numElements) {
 }
 
 
-__kernel void computeStatsFUNCTIONAL(__global double* buffer, ulong numElements) {
+__kernel void computeStatsFUNCTIONAL(__global double* buffer, int numElements) {
     size_t threadIdx = get_global_id(0);
 
     // Emulate a StatsAccumulator object
@@ -78,7 +78,7 @@ __kernel void computeStatsFUNCTIONAL(__global double* buffer, ulong numElements)
     double intPart = 0.0;
 
     // Process numElements elements
-    for (ulong i = 0; i < numElements; i += 1) {
+    for (int i = 0; i < numElements; i += 1) {
         double x = buffer[threadIdx*numElements+i];
         if (!valueNormalOrZero(x)) {
             // Skip if x is NaN, infinity or denormal
